crowd-risk-prediction/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0
fastapi>=0.85.0
uvicorn[standard]>=0.18.0
opencv-python>=4.6.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.1.0
scikit-image>=0.19.0
matplotlib>=3.5.0
pandas>=1.4.0
pyyaml>=6.0
tensorboard>=2.9.0
tqdm>=4.64.0
pillow>=9.0.0
pytest>=7.0.0
requests>=2.28.0
python-multipart>=0.0.5
pydantic>=1.10.0
albumentations>=1.3.0
seaborn>=0.11.0
plotly>=5.10.0
kaleido>=0.2.1